refactor: dedupe bitmap-font scripts into lib/bitmap-font
Remove duplicated scripts/lib/*.mjs modules and import shared logic from
lib/bitmap-font/*.ts via a Node strip-types resolve hook. Extend trace-core
and metrics-derive with script-only helpers previously duplicated in .mjs.
