+++ scripts/convert-bitmap-font.mjs
@@ -2,7 +2,7 @@
 import { mkdirSync, readFileSync, writeFileSync } from "fs";
 import { dirname, join } from "path";
 import { fileURLToPath } from "url";
-import { convertLegacyBitmapFont } from "./lib/convert-legacy-font.mjs";
+import { convertLegacyBitmapFont } from "../lib/bitmap-font/convert-legacy-font.ts";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const root = join(__dirname, "..");
