+++ scripts/lib/node-ts-hook.mjs
@@ -1,0 +1,4 @@
+import { register } from "node:module";
+import { pathToFileURL } from "node:url";
+
+register("./node-ts-resolve-hook.mjs", pathToFileURL("./scripts/lib/"));
